% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMBA-package.R
\docType{package}
\name{LSAMBA-package}
\alias{LSAMBA}
\alias{LSAMBA-package}
\title{LSAMBA: Lasso-SAMBA Algorithm}
\description{
Provides a lasso-based method for building mechanistic models using the SAMBA algorithm (Stochastic Approximation for Model Building Algorithm) (M Prague, M Lavielle (2022) \doi{10.1002/psp4.12742}). The package extends the 'Rsmlx' package (version 2024.1.0) to better handle high-dimensional data. It relies on the 'Monolix' software (version 2024R1; see (\url{https://monolixsuite.slp-software.com/monolix/2024R1/}), which must be installed beforehand.
}
\author{
\strong{Maintainer}: Auriane Gabaut \email{aurianegabaut@gmail.com}

Authors:
\itemize{
  \item Mélanie Prague
  \item Marc Lavielle (for former code of Rsmlx and connectors)
}

}
\keyword{internal}
